// src/client/app/_models/studyStates.ts
import { Injectable } from "@angular/core";

@Injectable()
export class StudyStates {
   
   idListaEstadoEstudio: number;
   codigoListaEstadoEstudio: string;
   nombreListaEstadoEstudio: string;
   ordenListaEstadoEstudio: number;
   indicadorHabilitado: boolean;
   auditoriaUsuario: number;
   auditoriaFecha: string;
   label: string;
   value: number;
   
}